app/requirements.txt
xlwt
django
pymysql
django-environ
django-rest-framework
django-ckeditor
django-crontab
psycopg2-binary
pyyaml
jsonfield
django-db-logger
drf-yasg
ua-parser 
user-agents
django-cleanup
DateTimeRange
pytz
django-cors-headers
django-auto-logout
cron-descriptor
requests
pyfcm
pandas
twilio
Pillow
stripe
qrcode
cryptography
weasyprint==52.2
xhtml2pdf
